(docs) : creating documentation for all endpoint using swager jsdoc (#5)
* (depen) : add swagger and express fileupload

* (feat) Users : adding endpoint update profile images

* (fix) Users : fix image url on register

* (feat) - Users : add middleware authentication

* (fix) - Users : middleware req variable

* (feat) - Users : add new endpoint login and create middleware

* (feat) - Users : adding endpoint to renew access token

* (feat) - Users : applying middleware for edit and delete user

* (feat) - Users : adding endpoint for logout

* (feat) - Projects : CRUD projects

* (readme) - adding readme

* (feat) - Projects : adding endpoint edit name, description, priority, deadline

* (feat) - Projects : adding create project with category and add more respone for get data

* (feat) - Tasks : creating endpoint CRUD task

* (fix) - Users : fixing user controller spacing

* (fix) - Projects : fixing endpoint get, and delete. and fix middleware

* (fix) - Projects : fixing all endpoint project update

* (feat) - Projects : adding endpoint update categories

* (feat) Task - Adding relation task and status

* (feat) - Project : add respone task in get projects

* (feat) - Tasks : change status, now use relational with table status

* (feat) - main : add route for task

* (feat) - Tasks : addjusting path api

* (refactor) - Project : change authorization to projectAuthor

* (delete) - Task : delete endpoint get all task

* (fix) - Task : fixing get task by id

* (fix) - Task : fixing update task statsu

* (fix) - Task : fixing delete task

* (refactor) : delete all console.log

* (fix) Users : Change respone for get all users

* (refactor) Users : Change respone for get user by id

* (refactor) Users : Change time maxage access token

* (docs) : setup swagger

* (docs) : creating docs for endpoint users

* (docs) : creating docs for endpoint task

* (docs) : creating docs for endpoint projects

* (docs) : instal swagger and cors types

* (feat) - Users : adding validation in delete account

* (refactor) - Project : change endpoint delete project

* (docs) : add endpoint delete user and change delete project

+++ src/controllers/Users.controllers.ts
@@ -27,7 +27,7 @@ export const getUsers = async (req: CustomRequest, res: Response, next: NextFunc
         updatedAt: true
       }
     })
-    res.status(200).json(users)
+    res.status(200).json({ users })
   } catch (error: any) {
     res.status(500).json(error.message)
   }
@@ -35,7 +35,7 @@ export const getUsers = async (req: CustomRequest, res: Response, next: NextFunc
 
 export const getUserById = async (req: CustomRequest, res: Response, next: NextFunction): Promise<any> => {
   try {
-    const users = await prisma.users.findUnique({
+    const user = await prisma.users.findUnique({
       where: { id: req.userId },
       select: {
         id: true,
@@ -46,7 +46,7 @@ export const getUserById = async (req: CustomRequest, res: Response, next: NextF
         updatedAt: true
       }
     })
-    res.status(200).json(users)
+    res.status(200).json({ user })
   } catch (error: any) {
     res.status(500).json(error.message)
   }
@@ -272,12 +272,20 @@ export const updateProfileImages = async (req: CustomRequest, res: Response, nex
 
 export const deleteUser = async (req: CustomRequest, res: Response, next: NextFunction): Promise<any> => {
   try {
-    const user = await prisma.users.delete({
+    const { password }: { password: string } = req.body
+    const user = await prisma.users.findUnique({
       where: { id: req.userId }
     })
     if (user == null) {
       return res.status(404).json({ message: 'User not found' })
     }
+    const passwordMatch = await argon2.verify(user.password, password)
+    if (!passwordMatch) {
+      return res.status(400).json({ msg: 'The password you entered is incorrect' })
+    }
+    await prisma.users.delete({
+      where: { id: user.id }
+    })
     res.status(200).json({ msg: 'Account deleted successfully' })
   } catch (error: any) {
     res.status(500).json({ msg: error.message })
@@ -305,7 +313,7 @@ export const login = async (req: CustomRequest, res: Response, next: NextFunctio
     const refreshToken = createRefreshToken(user.id)
     // Atur cookie access token
     res.cookie('access_token', accessToken, {
-      maxAge: 1200,
+      maxAge: 1200000,
       httpOnly: true
       // secure: true, // Hanya akan dikirim melalui HTTPS jika true
       // sameSite: "strict", // Atur ke 'lax' atau 'none' sesuai kebutuhan
